+++ user_management/api/urls/verify_email.py
@@ -8,8 +8,7 @@ urlpatterns = patterns(
     '',
     url(
         regex=(
-            r'^verify_email/(?P<uidb64>[0-9A-Za-z_\-]+)/'
-            r'(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/?$'
+            r'^verify_email/(?P<token>[0-9A-Za-z:\-_]+)/?$'
         ),
         view=csrf_exempt(views.VerifyAccountView.as_view()),
         name='verify_user',