+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## v10.1.0
+
+* Allow authenticated user to receive a new confirmation email.
+
+### Notes
+
+* Previously only anonymous could request a new confirmation email.
+
+## v10.0.0
+
+* Replace `default_token_generator` with `django.core.signing`.
+
+### Notes
+
+* Previously not validated emails would be invalid.
+
 ## v9.0.1
 
 * Send `user_logged_in` and `user_logged_out` signals from `GetAuthToken` view.